moved battery voltage functions to flipState lib
