+++ src/flipTurn-main.cpp
@@ -75,13 +75,11 @@
 #include <BleKeyboard.h>
 #include <Bounce2.h>
 
-#include "esp_adc_cal.h"  // Espressif Analog to Digital Converter (ADC) Calibration Driver library
-
 // internal (user) libraries:
-#include "controlRGB.h"   // rgb led control functions
 #include "myConstants.h"  // all constants in one file + pinout table
 #include "press_type.h"   // wrapper library further abstracting Yabl / Bounce2 switch routines
-#include "flipState.h"   //  library to manage flipTurn state machine
+#include "controlRGB.h"   // rgb led control functions
+#include "flipState.h"    //  library to manage flipTurn state machine
 
 //? ************** Selective Debug Scaffolding *********************
 // Selective debug scaffold: comment out  lines below to disable debugging tests at pre-processor stage
@@ -91,103 +89,16 @@
 #define DEBUG 1  // uncomment to debug
 //? ************ end Selective Debug Scaffolding ********************
 
-//   battery operating ranges
-#define HIGH_BATTERY_VOLTAGE 3.70  // 4.2 - 3.7V battery "fully" charged
-#define CHARGE_WARNING_TRIGGER_VOLTAGE 3.20 // trigger voltage to warn that device requires charging
-#define LOW_BATTERY_VOLTAGE 3.00   // lower bound battery operating range (DW01 battery protection circuit triggers at 2.4V )
-
-int current_battery_level = 100;  // initially set to fully charged, 100%
-
-const byte BLE_DELAY = 10;  // Delay (milliseconds) to prevent BT congestion
+extern const byte BLE_DELAY;  // Delay (milliseconds) to prevent BT congestion
+extern int current_battery_level;  // initially set to fully charged, 100%
 
 // blekeyboard instantiation params: (BT device name, BT Device manufacturer, Battery Level)
-BleKeyboard bleKeyboard("flipTurn", "CW Greenstreet", current_battery_level);
+//extern BleKeyboard bleKeyboard();
 
 // rgb led instantiation
 RgbLed rgbLed(RED_LED_PIN, GREEN_LED_PIN, BLUE_LED_PIN);
 
 bool hasRun = 0;  // run flag to control single execution within loop
-
-/*****************************************************************************
-Description : Reads the battery voltage through the voltage divider at AO pin (FireBeetle-ESP32, DFR0478)
-*!             Must solder-bridge zero-ohm pads to enable voltage divider hardware (ref DFR0478 Ver3 schematic)
-
-              Uses eFuse calibrations, if present (ESP32-E), otherwise alternative characterisation used
-              In comparison with a regular voltmeter, ESP32 vs. multimeter values differ only ~0.05V
-Input Value : -
-Return Value: battery voltage in volts
----------------------------------
-Ref:  ADC1_CHANNEL_0 Enumeration
-https://docs.espressif.com/projects/esp-idf/en/v4.1.1/api-reference/peripherals/adc.html#_CPPv414ADC1_CHANNEL_0)
-********************************************************************************/
-float readBattery() {
-    uint32_t value = 0;
-    int rounds = 11;
-    esp_adc_cal_characteristics_t adc_chars;
-
-    // battery voltage divided by 2 can be measured at GPIO36 (ADC1_CHANNEL0)
-    adc1_config_width(ADC_WIDTH_BIT_12);
-    adc1_config_channel_atten(ADC1_CHANNEL_0, ADC_ATTEN_DB_11);
-    switch (esp_adc_cal_characterize(ADC_UNIT_1, ADC_ATTEN_DB_11, ADC_WIDTH_BIT_12, 1100, &adc_chars)) {
-        case ESP_ADC_CAL_VAL_EFUSE_TP:
-#ifdef DEBUG
-            // Serial.println("Characterized using Two Point Value");
-#endif
-            break;
-        case ESP_ADC_CAL_VAL_EFUSE_VREF:
-#ifdef DEBUG
-            // Serial.printf("Characterised using eFuse Vref (%d mV)\r\n", adc_chars.vref);
-#endif
-            break;
-        default:
-            Serial.printf("Characterised using Default Vref (%d mV)\r\n", 1100);
-    }
-
-    // to avoid noise, sample the pin several times and average the result
-    for (int i = 1; i <= rounds; i++) {
-        value += adc1_get_raw(ADC1_CHANNEL_0);
-    }
-    value /= (uint32_t)rounds;
-
-    // due to the voltage divider (1M+1M), multiply value by 2 and convert mV to V
-    return esp_adc_cal_raw_to_voltage(value, &adc_chars) * 2.0 / 1000.0;
-}
-
-/*****************************************************************************
-Description : Tests for low battery charge (<=3V) and update BT Central device
-                with (%) battery charge level
-
-Input Value : battery_voltage (volts)
-Return Value: true / false
-********************************************************************************/
-bool isBatteryLow(uint32_t battery_voltage) {
-    static unsigned long updateTimer_msec = 0;
-
-    if (millis() - updateTimer_msec > 1000) {
-        bleKeyboard.setBatteryLevel(
-            battery_voltage >= HIGH_BATTERY_VOLTAGE ? 100 : 10 + 90 * (battery_voltage - LOW_BATTERY_VOLTAGE) / (HIGH_BATTERY_VOLTAGE - LOW_BATTERY_VOLTAGE));
-        delay(BLE_DELAY);
-        // Serial.printf("Battery: %d%%\n", 10 + 90 * (battery_voltage - LOW_BATTERY_VOLTAGE) / (HIGH_BATTERY_VOLTAGE - LOW_BATTERY_VOLTAGE));
-        updateTimer_msec = millis();
-    }
-
-    return battery_voltage <= LOW_BATTERY_VOLTAGE ? true : false;
-}
-
-/*****************************************************************************
-Description : calculates battery level (percentage) and updates BT Central device
-
-Input Value : battery voltage (in Volts)
-Return Value: -
-*******************************************************************************
-int setBatteryLevel(float battery_voltage) {
-    // do something
-    bleKeyboard.setBatteryLevel(current_battery_level);  // update battery level
-    {
-        // batteryAvg - LOW_VOLTAGE) / (HI_VOLTAGE - LOW_VOLTAGE)
-*/
-
-
 
 void setup() {
     Serial.begin(115200);
@@ -201,8 +112,6 @@ void setup() {
     // initialise button (eg switch) press_type set-up code (pin, pullup mode, callback function
     button.begin(SWITCH_PIN);
 
-    
-
 }  // end setup
 
 void loop() {
@@ -212,13 +121,13 @@ void loop() {
     /*
     float battery_voltage = readBattery();  // in Volts
     battery_voltage = 3.0;                  //! temporary debug line - remove!
-    */ 
+    */
 
     // TODO:  auto-shutdown if battery_voltage < 3V
 
     if (bleKeyboard.isConnected()) {
-        //rgbLed.setRgbColour(rgbLed.blue_BT_connected);
-        //delay(10);  //! temporary debug line.  Blocking!  remove
+        // rgbLed.setRgbColour(rgbLed.blue_BT_connected);
+        // delay(10);  //! temporary debug line.  Blocking!  remove
 
         //!  need to fix blink method - problem with arguments that are being passed
 
@@ -230,7 +139,8 @@ void loop() {
         }
 
         //! warning: delay() is blocking but necessary to prevent BT GATT overflow errors; must keep short or interferes with button presses
-        delay(BLE_DELAY);  //  optimised through trial & error, where no delay gives BT GATT overflow errors
+        //!  Currently, code will give linking error: "undefined reference to `BLE_DELAY'"
+        //delay(BLE_DELAY);  //  optimised through trial & error, where no delay gives BT GATT overflow errors
 
         if (button.triggered(SINGLE_TAP)) {
             yield();  // Do (almost) nothing - yield allows ESP8266 background functions
